implement wind sensor

+++ pypilot/autopilot.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-#   Copyright (C) 2016 Sean D'Epagnier
+#   Copyright (C) 2017 Sean D'Epagnier
 #
 # This Program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public
@@ -10,6 +10,7 @@
 # autopilot base handles reading from the imu (boatimu)
 
 import sys, getopt, os
+import math
 
 pypilot_dir = os.getenv('HOME') + '/.pypilot/'
 
@@ -20,11 +21,14 @@ from signalk.server import *
 from signalk.values import *
 from boatimu import BoatIMU
 
-import gpspoller
-import nmea_bridge
-import servo
-import math
+import gpspoller, wind, servo
 
+def resolv(angle, offset=0):
+    while offset - angle > 180:
+        angle += 360
+    while offset - angle < -180:
+        angle -= 360
+    return angle
 
 class Filter:
     def __init__(self,filtered, lowpass):
@@ -66,7 +70,6 @@ class AutopilotGain(RangeProperty):
       d['AutopilotGain'] = True
       return d
 
-
 class AutopilotBase(object):
   def __init__(self, *args, **keywords):
     super(AutopilotBase, self).__init__(*args, **keywords)
@@ -74,20 +77,24 @@ class AutopilotBase(object):
     self.boatimu = BoatIMU(self.server)
     self.servo = servo.Servo(self.server)
     self.gps = gpspoller.GpsPoller(self.server)
+    self.wind = wind.Wind(self.server)
 
     self.heading_command = self.Register(HeadingProperty, 'heading_command', 0)
     self.enabled = self.Register(BooleanProperty, 'enabled', False)
     self.mode = self.Register(EnumProperty, 'mode', 'compass', ['compass', 'gps', 'wind'], persistent=True)
     self.last_heading = False
 
+    self.heading = self.Register(Value, 'heading', 0)
     self.gps_heading_offset = self.Register(Value, 'gps_heading_offset', 0)
+    self.wind_heading_offset = self.Register(Value, 'wind_heading_offset', 0)
 
     # read initial value from imu as this takes time
-    while not self.boatimu.IMURead():
-        time.sleep(.1)
+#    while not self.boatimu.IMURead():
+#        time.sleep(.1)
 
   def __del__(self):
-      self.gps.process.terminate()
+      if self.gps.process:
+          self.gps.process.terminate()
       self.boatimu.__del__()
 
   def Register(self, _type, name, *args, **kwargs):
@@ -104,16 +111,16 @@ class AutopilotBase(object):
   def ap_iteration(self):
       period = .1
       data = False
-      while True:
+      for i in range(7):
           t0 = time.time()
           data = self.boatimu.IMURead()
           if data:
               break
-          time.sleep(period/2)
+          time.sleep(period/7)
               
       dt1 = time.time() - t0
 
-      if 'calupdate' in data and self.last_heading:
+      if data and 'calupdate' in data and self.last_heading:
           # with compass calibration updates, adjust the autopilot heading_command
           # to prevent actual course change
 
@@ -124,27 +131,43 @@ class AutopilotBase(object):
           while new_command < 0:
               new_command += 360
           self.heading_command.set(new_command)
+      if data:
+          self.last_heading = data['heading']
 
       dt2 = time.time() - t0
-      self.last_heading = data['heading']
 
       # calibration or other mode, disable autopilot
+      if not data:
+          #print 'no data!!!!!!!!!!!'
+          self.enabled.set(False)
+      
       if self.servo.drive.value == 'raw':
           self.mode.set('disabled')
 
       dt3 = time.time() - t0
-      magnetic_heading_command = self.heading_command.value
+
       self.gps.poll()
-      if self.gps.speed.value > 1:
-          track = self.gps.track.value
-          heading = self.boatimu.SensorValues['heading_lowpass'].value
-          d = .01
-          self.gps_heading_offset.set((1-d)*self.gps_heading_offset.value + d*(track - heading))
-          if self.mode.value == 'gps':
-              magnetic_heading_command += self.gps_heading_offset.value
+      self.wind.poll()
+      
+      compass_heading = self.boatimu.SensorValues['heading_lowpass'].value
+      if self.mode.value == 'compass':
+          self.heading.set(compass_heading)
+      elif self.mode.value == 'gps':
+        if self.gps.speed.value > 1:
+            offset = self.gps_heading_offset.value
+            diff = resolv(compass_heading - self.gps.track.value, offset)
+            d = .01
+            self.gps_heading_offset.set(resolv((1-d)*offset + d*diff))
+        self.heading.set(resolv(compass_heading - self.gps_heading_offset.value, 180))
+      elif self.mode.value == 'wind':
+        offset = self.wind_heading_offset.value
+        diff = resolv(compass_heading - self.wind.direction.value, offset)
+        d = .1
+        self.wind_heading_offset.set(resolv((1-d)*offset + d*diff))
+        self.heading.set(resolv(compass_heading - self.wind_heading_offset.value, 180))
 
       if self.enabled.value:
-          self.process_imu_data(self.boatimu, magnetic_heading_command)
+          self.process_imu_data(self.boatimu)
 
       t1 = time.time()
       if t1-t0 > period/2:
@@ -160,8 +183,9 @@ class AutopilotBase(object):
       t3 = time.time()
       if t3 - t2 > period/2:
           print 'gps is running too _slowly_', t2-t1
-          
-      self.server.HandleRequests(period - (t3 - t0))
+
+      t = max(period - (t3 - t0), .025)
+      self.server.HandleRequests(t)
 
 
 if __name__ == '__main__':
